# Performance improvement to unchecked segment ofNativeRestricted