@rsmogura 2021-01-16
Hi. I've added the benchmarks:
* benchmarks used by me;
* added a new benchmarks to `LoopOverNonConstant` to compare performance of this approach with existing methods for getting values.