@rsmogura 2021-01-16
> > Hi. I've added the benchmarks:
> 
> > 
> 
> >     * benchmarks used by me;
> 
> > 
> 
> >     * added a new benchmarks to `LoopOverNonConstant` to compare performance of this approach with existing methods for getting values.
> 
> 
> 
> It's strange to see the performances of global segment to be worse than that of a regular segment?

Indeed. I can't understand this (I'll make additional research). I suspect that's because the access is indexed (so global memory is treated as big array), or it's more about test itself - there's additional additions and we use longs instead of ints.