@mcimadamore 2021-01-18
>  could not find such optimization in case of global segment and long numbers.

The optimization happens on segment construction - if the segment size is < Integer.MAX_VALUE, the AbstractMemorySegment::defaultAccessMode method appends the SMALL flag to the set of segment flags. This flag is then consulted many times, typically before long addition/multiplications have to be performed.
> 
> From the other hand I have put together this PR and previous with Unsafe access, disabled alignment checks by settings alignment to 1, and global memory segment rocks. I added benchmark method to show performance of it.

Alignment check is an outstanding issue for the API - hotspot doesn't optimize those very well. See also:

https://mail.openjdk.java.net/pipermail/panama-dev/2021-January/011794.html