@mcimadamore 2021-01-18
Would you mind pasting the benchmark results before/after the patch, now that the benchmark has been fixed? Thanks!