@rsmogura 2021-01-18
After patch
```
Benchmark                                         Mode  Cnt  Score    Error  Units
LoopOverNonConstant.BB_get                        avgt   30  3.915 ?  0.012  ns/op
LoopOverNonConstant.BB_loop                       avgt   30  0.230 ?  0.001  ms/op
LoopOverNonConstant.global_segment_get            avgt   30  3.878 ?  0.001  ns/op
LoopOverNonConstant.global_segment_loop           avgt   30  0.400 ?  0.002  ms/op
LoopOverNonConstant.global_segment_loop_no_align  avgt   30  0.248 ?  0.001  ms/op
LoopOverNonConstant.segment_get                   avgt   30  5.497 ?  0.015  ns/op
LoopOverNonConstant.segment_loop                  avgt   30  0.229 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_readonly         avgt   30  0.239 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_slice            avgt   30  0.241 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_static           avgt   30  0.229 ?  0.001  ms/op
LoopOverNonConstant.unsafe_get                    avgt   30  3.434 ?  0.006  ns/op
LoopOverNonConstant.unsafe_loop                   avgt   30  0.229 ?  0.001  ms/op
```
before patch
```
Benchmark                                         Mode  Cnt  Score    Error  Units
LoopOverNonConstant.BB_get                        avgt   30  3.906 ?  0.012  ns/op
LoopOverNonConstant.BB_loop                       avgt   30  0.231 ?  0.001  ms/op
LoopOverNonConstant.global_segment_get            avgt   30  4.125 ?  0.012  ns/op
LoopOverNonConstant.global_segment_loop           avgt   30  0.596 ?  0.004  ms/op
LoopOverNonConstant.global_segment_loop_no_align  avgt   30  0.489 ?  0.002  ms/op
LoopOverNonConstant.segment_get                   avgt   30  5.507 ?  0.017  ns/op
LoopOverNonConstant.segment_loop                  avgt   30  0.231 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_readonly         avgt   30  0.238 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_slice            avgt   30  0.242 ?  0.001  ms/op
LoopOverNonConstant.segment_loop_static           avgt   30  0.230 ?  0.001  ms/op
LoopOverNonConstant.unsafe_get                    avgt   30  3.442 ?  0.008  ns/op
LoopOverNonConstant.unsafe_loop                   avgt   30  0.230 ?  0.001  ms/op

```